+++ packages/eslint-plugin-twofold/CHANGELOG.md
@@ -1,5 +1,7 @@
 # eslint-plugin-twofold
 
+## 0.0.19
+
 ## 0.0.18
 
 ## 0.0.17