+++ sites/website/CHANGELOG.md
@@ -1,5 +1,12 @@
 # website
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @twofold/framework@0.0.19
+
 ## 0.0.4
 
 ### Patch Changes